added favorites screen and made a separate dart files for categories/favorites
